# performance update kseq 