--- a/src/kseq.h
+++ b/src/kseq.h
@@ -10,7 +10,6 @@
 #define KS_SEP_TAB   1 // isspace() && !' '
 #define KS_SEP_LINE  2 // line separator: "\n" (Unix) or "\r\n" (Windows)
 #define KS_SEP_MAX   2
-
 #define BUF_SIZE     1048576
 
 #define ks_err(ks) ((ks)->end == -1)