as in [https://github.com/lh3/seqtk/pull/123](https://github.com/lmdu/pyfastx/commit/f746aa050aa7abd31d6d021270064a1b2ae86904)

Tested on fasta files speedup from 3.887 s ±  0.099 s  to 3.408 s ±  0.119 s